Rename dag_tree to tree.
